add session to friendTag route

+++ controllers/api/friendTag-route.js
@@ -1,16 +1,21 @@
 const router = require("express").Router();
 const { User, Music, FriendTag, MusicTag } = require("../../models/index.js");
-//path /
+//path :   /friendTag
 
-
-// Todo:
 // post route on friendTag when you click on following an user
+// user_id from currnet session
+// friend_id from html element
 router.post("/", async (req, res) => {
-  req.body = {
-    user_id: req.body.user_id,
-    friend_id: req.body.friend_id,
-  };
   try {
+    if (!req.session.logged_id) {
+      res.status(401).json("Please log in first!"); // 401 = Unauthorized error
+      console.log("the user is not logged in");
+      return;
+    }
+    req.body = {
+      user_id: req.session.user_id,
+      friend_id: req.body.friend_id,
+    };
     const newFriendTag = await FriendTag.create(req.body);
     res.status(200).json(newFriendTag);
     console.log("successfully added one friend tag!");
@@ -19,16 +24,18 @@ router.post("/", async (req, res) => {
   }
 });
 
-// Todo:
 // delete route when you click to unfollow an user
-// delete an friend tag of the logged in user, where the user_id and friend_id from comes from html element
-// the user_id must match the id of current user.
 router.delete("/", async (req, res) => {
   try {
+    if (!req.session.logged_id) {
+      res.status(401).json("Please log in first!"); // 401 = Unauthorized error
+      console.log("the user is not logged in");
+      return;
+    }
     const deleteFriendTag = await FriendTag.destroy({
       where: {
-        user_id: req.body.user_id,   // logged in user_id
-        friend_id: req.body.friend_id,  // from html element
+        user_id: req.session.user_id, // user_id from current session
+        friend_id: req.body.friend_id, // from html element
       },
     });
     if (!deleteFriendTag) {
